+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The [Unreleased](#unreleased) section collects notes for unreleased changes and 
 Note that not every tagged version may be suitable for production use. A Github
 release will be created for any release tested in production, and may be marked below with the tag [Release]
 
+## Unreleased
+
+- [Added] if block substitions
+
 ## [8.0.67] - 2022-05-17
 
 - [Fixed] migrations related to reference views